+++ lib/PhpParser/Node/Expr/FuncCall.php
@@ -16,8 +16,8 @@ class FuncCall extends Expr
      * Constructs a function call node.
      *
      * @param Node\Name|Expr $name       Function name
-     * @param Node\Arg[]                    $args       Arguments
-     * @param array                                   $attributes Additional attributes
+     * @param Node\Arg[]     $args       Arguments
+     * @param array          $attributes Additional attributes
      */
     public function __construct($name, array $args = [], array $attributes = []) {
         parent::__construct($attributes);