+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-Version 4.0.1-dev
+Version 4.0.2-dev
 -----------------
+
+Nothing yet.
+
+Version 4.0.1 (2018-03-25)
+--------------------------
 
 ### Added